[ENG-1103] Generate Leap’s C# SDK (#228)
* fix formatting of array type parameter for form data

* improve README for C# SDK

* improve formatting of templates

TODO: generate tests for C# SDK

* override AbstractCSharpCodegen

* render double for all "number" type schemas

* improve testing + docs in C# SDK

* implement debugging of native java classes

* support parent debugging

* fix formatting of Array in C# SDK

* add changeset

* update snapshot

+++ generator/konfig-generator-api/src/main/java/org/openapitools/codegen/languages/CSharpNetCoreClientCodegen.java
@@ -152,7 +152,7 @@ public class CSharpNetCoreClientCodegen extends AbstractCSharpCodegen {
         typeMapping.put("float", "float");
         typeMapping.put("long", "long");
         typeMapping.put("double", "double");
-        typeMapping.put("number", "decimal");
+        typeMapping.put("number", "double");
         typeMapping.put("decimal", "decimal");
         typeMapping.put("DateTime", "DateTime");
         typeMapping.put("date", "DateTime");
